+++ src/environments/environment.prod.ts
@@ -7,4 +7,7 @@ export const environment = {
   oidcAuthorizationEndpoint: window["env"]["oidc_authorization_endpoint"],
   oidcClientId: window["env"]["oidc_client_id"],
   oidcPortalRedirectUri: window["env"]["oidc_portal_redirect_uri"],
+  issuanceEntryPoint: window["env"]["issuance_entry_point"] ?? '',
+  entryPointTargetWithValidation: window["env"]["entry_point_target_with_validation"] ?? '',
+  entryPointTargetDirect: window["env"]["entry_point_target_direct"] ?? '',
 };